Break out previous matchups component
