+++ components/Game/index.ts
@@ -1,5 +1,6 @@
 import DivisionStandings from './Preview/DivisionStandings';
 import GoalieComparison from './Preview/GoalieComparison';
+import PreviousMatchups from './Preview/PreviousMatchups';
 import SkaterComparison from './Preview/SkaterComparison';
 import TeamsBlock from './Preview/TeamsBlock';
 import TeamStats from './Preview/TeamStats';
@@ -7,6 +8,7 @@ import TeamStats from './Preview/TeamStats';
 export {
   DivisionStandings,
   GoalieComparison,
+  PreviousMatchups,
   SkaterComparison,
   TeamsBlock,
   TeamStats
