+++ pages/[league]/[season]/game/[gameId].tsx
@@ -9,10 +9,10 @@ import styled from 'styled-components';
 import { PulseLoader } from 'react-spinners';
 
 import Header from '../../../../components/Header';
-import { DivisionStandings, GoalieComparison, SkaterComparison, TeamsBlock, TeamStats } from '../../../../components/Game';
+import { DivisionStandings, GoalieComparison, PreviousMatchups, SkaterComparison, TeamsBlock, TeamStats } from '../../../../components/Game';
 import { Matchup as MatchupData } from '../../../api/v1/schedule/game/[gameId]';
 import { Standings } from '../../../api/v1/standings';
-import { FlexColumn, SectionTitle, TeamLogoSmall } from '../../../../components/Game/common';
+import { FlexColumn, } from '../../../../components/Game/common';
 
 interface Props {
   league: string;
@@ -66,28 +66,6 @@ function GameResults({ league, gameId }: Props): JSX.Element {
     setDivisions([awayDivision, homeDivision]);
   }, [divisionData, gameData, divisions]);
 
-  const renderPreviousMatchups = () => gameData.previousMatchups.map((matchup) => (
-    <Matchup key={matchup.slug}>
-      <SectionTitle>
-        {matchup.date}
-      </SectionTitle>
-      <MatchupTeamRow>
-        <TeamLogoSmall>
-          {matchup.awayTeam === gameData.game.awayTeam ? <Sprites.Away /> : <Sprites.Home />}
-        </TeamLogoSmall>
-        <span>{matchup.awayTeam === gameData.game.awayTeam ? gameData.teams.away.nickname : gameData.teams.home.nickname}</span>
-        <MatchupRowScore lost={matchup.awayScore < matchup.homeScore}>{matchup.awayScore}</MatchupRowScore>
-      </MatchupTeamRow>
-      <MatchupTeamRow>
-        <TeamLogoSmall>
-          {matchup.homeTeam === gameData.game.homeTeam ? <Sprites.Home /> : <Sprites.Away />}
-        </TeamLogoSmall>
-        <span>{matchup.homeTeam === gameData.game.homeTeam ? gameData.teams.home.nickname : gameData.teams.away.nickname}</span>
-        <MatchupRowScore lost={matchup.homeScore < matchup.awayScore}>{matchup.homeScore}</MatchupRowScore>
-      </MatchupTeamRow>
-    </Matchup>
-  ));
-
   const isLoading = isLoadingAssets || !gameData;
   const isRegularSeason = gameData && gameData.game.type === "Regular Season";
 
@@ -107,6 +85,7 @@ function GameResults({ league, gameId }: Props): JSX.Element {
           </ErrorBlock>
         }
         {isLoading && !gameError && <PulseLoader size={15} />}
+
         {!isLoading && (
           <>
             <FlexColumn width={300}>
@@ -123,21 +102,7 @@ function GameResults({ league, gameId }: Props): JSX.Element {
               <SkaterComparison gameData={gameData} Sprites={Sprites} />
               <GoalieComparison gameData={gameData} Sprites={Sprites} />
             </Comparison>
-            <PreviousMatchups>
-              <MatchupsHeader>
-                <SectionTitle>
-                  {`${gameData.game.type} Series`}
-                </SectionTitle>
-              </MatchupsHeader>
-              {gameData.previousMatchups.length > 0 && renderPreviousMatchups()}
-              {gameData.previousMatchups.length === 0 && (
-                <Matchup>
-                  <div>
-                    No previous games played
-                  </div>
-                </Matchup>
-              )}
-            </PreviousMatchups>
+            <PreviousMatchups gameData={gameData} Sprites={Sprites} />
           </>
         )}
       </Container>
@@ -172,67 +137,6 @@ const Comparison = styled.div`
 `;
 
 // Right
-const PreviousMatchups = styled.div`
-  display: flex;
-  flex-direction: column;
-  background-color: ${({ theme }) => theme.colors.grey100};
-  padding: 0 15px;
-  width: 300px;
-
-  div:last-child {
-    margin-bottom: 0;
-    border-bottom: none;
-  }
-`;
-
-const MatchupTeamRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: center;
-  font-family: Montserrat, sans-serif;
-  font-size: 14px;
-  font-weight: 600;
-
-  div:first-child {
-    margin-right: 10px;
-  }
-`;
-
-const MatchupRowScore = styled.span<{
-  lost?: boolean;
-}>`
-  ${({ lost, theme }) => lost && `color: ${theme.colors.grey500};`};
-  font-family: Montserrat, sans-serif;
-  margin-left: auto;
-  font-size: 16px;
-`;
-
-const MatchupsHeader = styled.div`
-  font-weight: 600;
-  border-bottom: 2px solid ${({ theme }) => theme.colors.grey300};
-  margin-bottom: 10px;
-  padding: 10px 0;
-`;
-
-const Matchup = styled.div`
-  display: flex;
-  flex-direction: column;
-  padding-bottom: 15px;
-  margin-bottom: 15px;
-  border-bottom: 2px solid ${({ theme }) => theme.colors.grey300};
-
-  * {
-    padding-bottom: 5px;
-  }
-
-  span:first-child {
-    font-family: Montserrat, sans-serif;
-    font-size: 14px;
-    margin-bottom: 10px;
-  }
-`;
-
 // Error
 const ErrorBlock = styled.div`
   display: flex;
